refactor(#3276): decompose compilePropertyAccess slice 4 — extract terminal tail (COMPLETES)
Wave B (subtask of #3182), stacked on slice 3. Extract the terminal struct-name
resolution + dynamic member-get tail of compilePropertyAccess into
finalizeStructAndDynamicMemberGet. The region always returns, so it is a plain
tail-extraction (return finalize(...), no PA_FALLTHROUGH sentinel).

property-access.ts: 5573 -> 4773 LOC (-800). compilePropertyAccess itself
~3335 -> 96 LOC (preamble + 15 guard-band call blocks + one terminal return).
Cumulative across all 4 slices: 7989 -> 4773 LOC (-3216); 16 named helpers now
in property-access-dispatch.ts.

Byte-identity: IDENTICAL 39/39 (gc/standalone/wasi). tsc 0. Smoke 13/13.

✓
